# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlowBook - Multi-vendor Hair Salon Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/favorites:
    get:
      operationId: getFavorites
      tags: [users]
      summary: Get user's favorite salons
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorite salons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Salon"
  /users/favorites/{salonId}:
    post:
      operationId: addFavorite
      tags: [users]
      summary: Add salon to favorites
      security:
        - bearerAuth: []
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    delete:
      operationId: removeFavorite
      tags: [users]
      summary: Remove salon from favorites
      security:
        - bearerAuth: []
      parameters:
        - name: salonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── CATEGORIES ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, owner, admin]
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    # ─── CATEGORY ─────────────────────────────────────────────────────────
...
    Salon:
      type: object
      required: [id, name, ownerId, address, city, isActive, avgRating, totalReviews, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        ownerId:
          type: integer
        address:
          type: string
        city:
          type: string
        state:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        avgRating:
          type: number
        totalReviews:
          type: integer
        isActive:
          type: boolean
        isVerified:
          type: boolean
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        distanceKm:
          type: ["number", "null"]
        isFavorited:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
...